Question analysis, basic structure, and testing of sorting before writing function

+++ src/com/codewithdelayne/MiniMaxSum.java
@@ -5,8 +5,8 @@ public class MiniMaxSum {
     static void miniMaxSum(int[] arr) {
 
         Arrays.sort(arr);
-        int min = 0;
-        int max = 0;
+        long min = 0;
+        long max = 0;
 
         for(int i = 0; i < arr.length - 1; i++){
             min = min + arr[i];
